requirements.txt
aws-cdk-lib==2.178.1
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
aws-lambda-powertools>=2.0.0
python-jose>=3.3.0
aws-xray-sdk>=2.12.0